show order

+++ database/migrations/2025_05_01_070529_create_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->string('order_number')->unique();
             $table->foreignId('user_id')->nullable()->constrained('users')->nullOnDelete();
             $table->float('discount')->nullable();
-            $table->float('total');
+            $table->double('total', 8, 2);
             $table->string('payment_method');
             $table->text('notes')->nullable();
             $table->enum('payment_status', \App\Enums\PaymentStatus::values())->default('غير مدفوع');
